test(cluster/brokers cli) update clusters and brokers tests
